+++ Src/Classes/BuffDefinition.lua
@@ -109,6 +109,8 @@ end
 ---@field singleLink string Printable link for single buff. Use buffdef:SingleLink() to safely handle missing value
 ---@field singleMana number Mana cost
 ---@field singleText string Name of single buff spell (from GetSpellInfo())
+---@field consumeGroupTitle string Override spell name for groups of consumable buffs providing same stat
+---@field consumeGroupIcon WowIconId Override buff icon for groups of consumables
 ---@field skipList string[] If spell cast failed, contains recently failed targets
 ---@field spellIcon WowIconId
 ---@field targetClasses BomClassName[] List of target classes which are shown as toggle boxes to enable cast per class
@@ -164,6 +166,20 @@ function buffDefModule:genericConsumable(allBuffs, singleId, providedByItem)
                          :IsConsumable(true)
                          :IsDefault(false)
                          :CreatesOrProvidedByItem(providedByItem)
+  return b
+end
+
+---@param allBuffs BomBuffDefinition[]
+---@param buffId NewBuffId
+---@param providesAuras WowSpellId[] Auras from the food
+---@param providedByItems WowItemId[] Item or multiple items giving this buff
+---@return BomBuffDefinition
+function buffDefModule:consumableGroup(allBuffs, buffId, providesAuras, providedByItems)
+  local b = buffDefModule:createAndRegisterBuff(allBuffs, providesAuras[1], nil)
+                         :IsConsumable(true)
+                         :IsDefault(false)
+                         :ProvidesAuras(providesAuras)
+                         :CreatesOrProvidedByItem(providedByItems)
   return b
 end
 
@@ -296,6 +312,13 @@ function buffDefClass:IsConsumable(isConsum)
   return self
 end
 
+local function reverseTable(tab)
+  for i = 1, math.floor(#tab / 2), 1 do
+    tab[i], tab[#tab - i + 1] = tab[#tab - i + 1], tab[i]
+  end
+  return tab
+end
+
 ---@param itemId WowItemId|WowItemId[]
 ---@return BomBuffDefinition
 function buffDefClass:CreatesOrProvidedByItem(itemId)
@@ -304,6 +327,13 @@ function buffDefClass:CreatesOrProvidedByItem(itemId)
   else
     self.items = --[[---@type WowItemId[] ]] itemId
   end
+
+  -- Clone item ids, and reverse
+  self.itemsReverse = {}
+  for _, val in pairs(--[[---@not nil]] self.items) do
+    table.insert(self.itemsReverse, val)
+  end
+  reverseTable(self.itemsReverse)
 
   return self
 end
@@ -455,6 +485,15 @@ end
 ---@return BomBuffDefinition
 function buffDefClass:GroupDuration(duration)
   self.groupDuration = duration
+  return self
+end
+
+---@param title string
+---@param icon WowIconId
+---@return BomBuffDefinition
+function buffDefClass:BuffTitle(title, icon)
+  self.consumeGroupTitle = title
+  self.consumeGroupIcon = icon
   return self
 end
 
@@ -675,7 +714,7 @@ end
 
 ---Call function with the icon when icon value is ready, or immediately if value
 ---is available. This allows for late loaded icons.
----@param iconReadyFn fun(icon: string)
+---@param iconReadyFn fun(icon: string|WowSpellId)
 function buffDefClass:GetIcon(iconReadyFn)
   if self.itemIcon then
     iconReadyFn(self.itemIcon) -- value was ready
@@ -705,7 +744,7 @@ end
 
 function buffDefClass:IsItem()
   -- TODO: self.isConsumable does this too?
-  return self.items and next(self.items) ~= nil
+  return self.items and next(--[[---@not nil]] self.items) ~= nil
 end
 
 ---@param unit BomUnit
@@ -769,12 +808,6 @@ function buffDefClass:RefreshTextAndIcon(iconReadyFn, nameReadyFn)
   -- nil otherwise
 end
 
------@return WowSpellId
---function buffDefClass:GetFirstSingleId()
---  local _, singleId = next(self.singleFamily)
---  return singleId
---end
-
 ---@return WowItemId|nil
 function buffDefClass:GetFirstItem()
   local _, itemId = next(self.items)
@@ -810,7 +843,14 @@ function buffDefClass:GetDownRank(spellId)
   return downrank -- unsuccessful but return whatever found
 end
 
-function buffDefClass:SingleLink()
+---@param bestItemIdAvailable WowItemId|nil If set, will request item link to a specific item
+function buffDefClass:SingleLink(bestItemIdAvailable)
+  if bestItemIdAvailable then
+    local itemInfo = BOM.GetItemInfo(--[[---@not nil]] bestItemIdAvailable)
+    if itemInfo then
+      return (--[[---@not nil]] itemInfo).itemLink
+    end
+  end
   return (self.singleLink or self.singleText) or "?"
 end